Revert "Revert "Revert "change deprecated tf.to_float to tf.cast"""
This reverts commit 2846b3cf4aaf2adfa69b8a6c673e661084fcb1b0.

+++ models-r1.13.0/official/transformer/model/model_utils.py
@@ -42,16 +42,13 @@ def get_position_encoding(
   Returns:
     Tensor with shape [length, hidden_size]
   """
-  # position = tf.to_float(tf.range(length))
-  position = tf.cast(tf.range(length), tf.float32)
+  position = tf.to_float(tf.range(length))
   num_timescales = hidden_size // 2
   log_timescale_increment = (
       math.log(float(max_timescale) / float(min_timescale)) /
-      (tf.cast(num_timescales, tf.float32) - 1))
-      # (tf.to_float(num_timescales) - 1))
+      (tf.to_float(num_timescales) - 1))
   inv_timescales = min_timescale * tf.exp(
-      tf.cast(tf.range(num_timescales), tf.float32) * -log_timescale_increment)
-      # tf.to_float(tf.range(num_timescales)) * -log_timescale_increment)
+      tf.to_float(tf.range(num_timescales)) * -log_timescale_increment)
   scaled_time = tf.expand_dims(position, 1) * tf.expand_dims(inv_timescales, 0)
   signal = tf.concat([tf.sin(scaled_time), tf.cos(scaled_time)], axis=1)
   return signal
@@ -89,8 +86,7 @@ def get_padding(x, padding_value=0):
       0 -> non-padding, 1 -> padding
   """
   with tf.name_scope("padding"):
-    # return tf.to_float(tf.equal(x, padding_value))
-    return tf.cast(tf.equal(x, padding_value), tf.float32)
+    return tf.to_float(tf.equal(x, padding_value))
 
 
 def get_padding_bias(x):
